Fix the bug of compile model to tensorrt
Closes #491
