+++ model_compiler/pytest.ini
@@ -1,2 +1,4 @@
 [pytest]
-markers = gpu_test
+markers =
+    gpu_test
+    no_test
