+++ model_compiler/setup.py
@@ -56,9 +56,9 @@ setuptools.setup(
         'future',
         'networkx',
         'tensorflow==2.4.0',
-        'torch==1.7.1',
+        'torch',
         'onnx-tf',
-        'onnx-caffe2==1.0.0',
+        'onnx-caffe2',
         'paddlepaddle',
         'paddle2onnx',
         'tensorflow_addons',
