+++ model_compiler/src/model_compiler/compilers/torch_model_file_to_onnx_model.py
@@ -14,25 +14,7 @@ from .. import utilities
 from ..models.sources.torch_model_file import TorchModelFile
 from ..models.irs.onnx_model import OnnxModel
 from ..models.data_format import DataFormat
-
-
-# If put this function in data_type.py, pytest has bug about "import torch"
-def from_torch_data_type(type_str):
-    torch_data_type_map = {
-        'FLOAT': torch.float,
-        'DOUBLE': torch.double,
-        'COMPLEX64': torch.complex64,
-        'COMPLEX128': torch.complex128,
-        'FLOAT16': torch.float16,
-        'BFLOAT16': torch.bfloat16,
-        'UINT8': torch.uint8,
-        'INT8': torch.int8,
-        'INT16': torch.int16,
-        'INT32': torch.int32,
-        'INT64': torch.int64,
-        'BOOL': torch.bool
-    }
-    return torch_data_type_map[type_str.upper()]
+from ..models.data_type import DataType
 
 
 class Config(NamedTuple):
@@ -46,14 +28,14 @@ class Config(NamedTuple):
     def from_json(value: Mapping[str, Any]) -> 'Config':
         return Config(input_names=value['input_names'],
                       input_shapes=utilities.get_input_shapes(value.get('input_shapes')),
-                      data_type=from_torch_data_type(value['data_type']),
+                      data_type=DataType.from_torch_data_type(value['data_type']),
                       max_batch_size=value['max_batch_size'],
                       input_formats=utilities.get_data_formats(value.get('input_formats')))
 
     @staticmethod
     def from_env(env: Mapping[str, str]) -> 'Config':
         input_shapes = utilities.get_input_shapes_from_env(env.get('INPUT_SHAPES'))
-        data_type = from_torch_data_type(env.get('DATA_TYPE'))
+        data_type = DataType.from_torch_data_type(env.get('DATA_TYPE'))
 
         return Config(input_names=env['INPUT_NAMES'].split(','),
                       input_shapes=input_shapes,
