+++ model_compiler/tox.ini
@@ -11,7 +11,9 @@ commands = {envpython} tools/install-tensorrt.py
            mypy --install-types --non-interactive -p model_compiler -p tests.model_compiler
            pylint src tests
 
-           cpu: pytest -m 'not gpu_test' --cov --cov-config .coveragerc-cpu -n auto
+           {envpython} setup.py build_py
+
+           cpu: pytest -m 'not gpu_test and not no_test' --cov --cov-config .coveragerc-cpu -n auto
            gpu: pytest --cov -n auto
 
 deps = .[test]
